로그인 및 회원정보 접근권한 등... 기능 추가

+++ src/main/java/com/prj1/controller/BoardController.java
@@ -3,6 +3,7 @@ package com.prj1.controller;
 import com.prj1.domain.Board;
 import com.prj1.service.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,10 +23,8 @@ public class BoardController {
     }
 
     @PostMapping("/add")
-    public String addPost(Board board, RedirectAttributes rttr) {
-        System.out.println("board = " + board);
-
-        service.add(board);
+    public String addPost(Board board, Authentication authentication, RedirectAttributes rttr) {
+        service.add(board, authentication);
 
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
@@ -55,8 +54,10 @@ public class BoardController {
     }
 
     @PostMapping("/delete")
-    public String delete(Integer id) {
-        service.remove(id);
+    public String delete(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
 
         return "redirect:/";
     }
@@ -71,10 +72,10 @@ public class BoardController {
     }
 
     @PostMapping("/modify")
-    public String modifyPost(Board board, RedirectAttributes rttr) {
-        System.out.println("board = " + board);
-        service.modify(board);
-
+    public String modifyPost(Board board, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(board.getId(), authentication)) {
+            service.modify(board);
+        }
         rttr.addAttribute("id", board.getId());
         return "redirect:/board";
     }
