+++ src/main/java/com/prj1/controller/MemberController.java
@@ -3,6 +3,7 @@ package com.prj1.controller;
 import com.prj1.domain.Member;
 import com.prj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -44,10 +45,11 @@ public class MemberController {
     }
 
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
-
-        return "redirect:/member/list";
+    public String remove(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
+        return "redirect:/logout";
     }
 
     @GetMapping("modify")
@@ -58,9 +60,10 @@ public class MemberController {
     }
 
     @PostMapping("modify")
-    public String modify(Member member, RedirectAttributes rttr) {
-        service.modify(member);
-
+    public String modify(Member member, Authentication authentication, RedirectAttributes rttr) {
+        if (service.hasAccess(member.getId(), authentication)) {
+            service.modify(member);
+        }
         rttr.addAttribute("id", member.getId());
         return "redirect:/member";
     }
@@ -79,6 +82,11 @@ public class MemberController {
         int check = service.nickNameCheck(nickName);
 
         return check;
+    }
+
+    @GetMapping("login")
+    public String loginForm() {
+        return "member/login";
     }
 
 }
